5 New Evo Methods
Thanks to pfero, here's their summary of this commit:

"Added a few new evolution methods. Some a little more useful than the
others.
- EVOLVE_MOVE: Causes the 'mon to evolve when it is leveled up while knowing a
certain move.
- EVOLVE_MOVE_TYPE: Causes the 'mon to evolve when it is leveled up while
knowing a move of a certain type.
- EVOLVE_ITEM_GENDER: Causes the 'mon to evolve when an item is used and it
has the correct gender.
- EVOLVE_HOLD: Causes the 'mon to evolve when it's leveled up while holding an
item during a specific time of day.
- EVOLVE_PARTY: Causes the 'mon to evolve when it's leveled up with another
specific 'mon in the party.

Also reordered the entire EvolveAfterBattle_MasterLoop to be easier to
navigate and adapt.
Also rewrote and fixed a lot of possible corner cases in
.DetectCompatibility (engine/party_menu.asm). It's slightly slower, but
it'll allow more than 3 item evolves per mon and any possible order of
evolution checks. To accomplish this, I added EvoTableSizes to
data/evos_attacks.asm and SkipEvo to home/evolve.asm."

It's an old commit, so labels had to be updated, but other than that it built fine. The only method I actually tested was the EVOLVE_PARTY, as I may not need the others. It seemed to work as intended!