Bridge Encounters /Improved Outdoor Sprite System
Started hard coding encounters for Wingull/Yanma on the Seabite Bridge, very similar to Driftveil Bridge. If you interact with a cloud of feathers, you'll either face a wild pokemon, get an item, or nothing will happen. Haven't put in items yet.

Also implemented the tutorial for improving the outdoor sprite system:

https://github.com/pret/pokecrystal/wiki/Improve-the-outdoor-sprite-system

because the game did not want to cooperate with my new FEATHERS sprite otherwise